simplify sweep
