+++ cli/commands/test.ts
@@ -174,18 +174,12 @@ async function runTestSuites(
 
   const suites: RunbookTestSuite[] = []
 
-  // Run parallel suites
-  if (parallel.length > 0) {
-    // For simplicity, run sequentially for now.
-    // True parallelism would require worker_threads or child_process forking.
-    // TODO: Add parallel execution with worker pool
-    for (const runbook of parallel) {
-      suites.push(await runTestSuite(runbook, opts))
-    }
-  }
-
-  // Run sequential suites
-  for (const runbook of sequential) {
+  // Parallelizable runbooks currently run sequentially as well — true
+  // parallelism would require a worker pool (worker_threads / child_process).
+  // Until then both groups share one loop, parallelizable first to preserve
+  // report ordering.
+  // TODO: Add parallel execution with worker pool for the `parallel` group.
+  for (const runbook of [...parallel, ...sequential]) {
     suites.push(await runTestSuite(runbook, opts))
   }
